backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
httpx==0.27.2
numpy==1.26.4
pandas==2.2.3
cohere==5.11.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
